backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.1",
    "fastapi>=0.135.1",
    "firebase-admin>=6.8.0",
    "motor>=3.7.1",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx==0.27.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
